+++ src/components/title1/Title1ProgramTable.tsx
@@ -593,7 +593,7 @@ function Title1ProgramTable({
                                 >
                                     <i>
                                         The payments,base acres and payment recipients are calculated as the total of
-                                        the data from 2014-2021. 2022 payments for Title I have not yet been paid.
+                                        the data from 2014-2023. 2024 payments for Title I have not yet been paid.
                                     </i>
                                 </Typography>
                             ) : (