+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.11.0] - 2025-07-31
 
 ### Added
 - Add the 'Download CSV' feature to all tables in Title I page [#412](https://github.com/policy-design-lab/pdl-frontend/issues/412)
@@ -387,6 +387,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Map data json [#12](https://github.com/policy-design-lab/pdl-frontend/issues/12)
 - Final landing page changes for initial milestone [#15](https://github.com/policy-design-lab/pdl-frontend/issues/15)
 
+[1.11.0]: https://github.com/policy-design-lab/pdl-frontend/compare/1.10.0...1.11.0
 [1.10.0]: https://github.com/policy-design-lab/pdl-frontend/compare/1.9.0...1.10.0
 [1.9.0]: https://github.com/policy-design-lab/pdl-frontend/compare/1.8.0...1.9.0
 [1.8.0]: https://github.com/policy-design-lab/pdl-frontend/compare/1.7.0...1.8.0